keep pool generation id 32-bit
will likely go back to 32-bit in total with bit packing later

+++ include/barelymusician.h
@@ -420,7 +420,7 @@ typedef struct BarelyRef {
   uint32_t index;
 
   /// Generation.
-  uint16_t generation;
+  uint32_t generation;
 } BarelyRef;
 
 /// Instrument control override.
